brain/requirements.txt
# Web Framework
fastapi
uvicorn

# Database & Cache
redis
python-dotenv

# Machine Learning & Data Processing
scikit-learn
xgboost
pandas
joblib
psycopg2-binary
SQLAlchemy
geoalchemy2
networkx
osmnx
geopandas